+++ cmd/hidden_lake/helpers/traffic/internal/handler/service_tcp.go
@@ -33,8 +33,7 @@ func HandleServiceTCP(pCfg config.IConfig, pDBWrapper database.IDBWrapper, pLogg
 			WithConn(pConn).
 			WithHash(pNetMsg.GetHash()).
 			WithProof(pNetMsg.GetProof()).
-			WithSize(len(pld.GetBody())).
-			WithRecv(true)
+			WithSize(len(pld.GetBody()))
 
 		if _, err := message.LoadMessage(pCfg.GetSettings(), pld.GetBody()); err != nil {
 			pLogger.PushWarn(logBuilder.WithType(anon_logger.CLogWarnMessageNull))