+++ cmd/hidden_lake/service/internal/handler/service_tcp.go
@@ -30,8 +30,7 @@ func HandleServiceTCP(pMutex *sync.Mutex, pCfgW config.IWrapper, pLogger logger.
 		// enrich logger
 		logBuilder.
 			WithSize(len(reqBytes)).
-			WithPubKey(sender).
-			WithRecv(true)
+			WithPubKey(sender)
 
 		cfg := pCfgW.GetConfig()
 		friends := cfg.GetFriends()