+++ internal/logger/anon/anon_test.go
@@ -50,7 +50,7 @@ func (p *tsAddr) String() string  { return "192.168.0.1:2000" }
 const (
 	tcService = "TST"
 	tcHash    = "96cb1f0968adba001ebc216708a02c8d2817b1a77fad1206012c22716a9b130b"
-	tcFmtLog  = "service=TST path=RECV type=ENQRQ hash=96CB1F09...6A9B130B addr=6245E00D...327047E9 proof=0000012345 size=1024B conn=192.168.0.1:2000"
+	tcFmtLog  = "service=TST type=ENQRQ hash=96CB1F09...6A9B130B addr=6245E00D...327047E9 proof=0000012345 size=1024B conn=192.168.0.1:2000"
 )
 
 func TestLoggerPanic(t *testing.T) {
@@ -100,6 +100,5 @@ func testNewAnonLogger() anon_logger.ILogBuilder {
 		WithProof(12345).
 		WithSize(1024).
 		WithPubKey(asymmetric.LoadRSAPubKey(testutils.TgPubKeys[0])).
-		WithConn(&tsConn{}).
-		WithRecv(true)
+		WithConn(&tsConn{})
 }