+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### CHANGES
 
 - Update `pkg/logger`: *os.File -> io.Writer
+- Update `pkg/network/anonymity/logger`: delete GetRecv, WithRecv methods
+- Update `cmd/hidden_lake/*/_daemon`: delete from status timestamp systemd
 
 <!-- ... -->
 
@@ -37,7 +39,6 @@
 - Update `pkg/network/anonymity`: rename BroadcastPayload -> SendPayload
 - Update `pkg/network/anonymity/examples`: the examples is simplified
 - Update `pkg/logger`: if return string is nil -> log is not printing
-- Update `pkg/network/anonymity/logger`: append GetRecv, WithRecv methods
 
 ### BUG FIXES